freebsd-support: PR changes and fixes
* Functions in fusefrontend_reverse/node_xattr_freebsd.go have been
  stubbed for now.

* asuser_freebsd.go updated to only run f() when context is nil;
  otherwise log a warning and return an error.

* emulate.go build flags updated, and FreeBSD specific version added.

* sys_freebsd.go bug in Renameat2 with RENAME_EXCHANGE flag fixed.
  FreeBSD does not support atomic file swapping, so this flag now
  returns an error.

* unix2syscall and atime is identical between FreeBSD and Darwin,
  updated filenames so Go will build the file for FreeBSD and Mac OS.

+++ internal/syscallcompat/emulate_freebsd.go
@@ -1,5 +1,3 @@
-//go:build !freebsd
-
 package syscallcompat
 
 import (
@@ -27,5 +25,5 @@ func emulateMknodat(dirfd int, path string, mode uint32, dev int) error {
 		}
 		defer syscall.Fchdir(cwd)
 	}
-	return syscall.Mknod(path, mode, dev)
+	return syscall.Mknod(path, mode, uint64(dev))
 }
